requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-genai==1.10.0
google-api-python-client==2.143.0
google-auth==2.34.0
pydantic==2.8.2
python-dotenv==1.0.1
